added recursive function
traversal is a recursive implementation of code which performs similar traversal below. This means that there's two traversals going on - a recursive one, and a set of loops. 

The loops iterate over the entire dataset, and the recursive starts from the outlet and "marches"

The loops are more mature and have comparisons, a "visited" dict, etc. This is needed in the recursive function as an exit condition. Currently there is no exit condition, so don't run this code!!