adding merged code
this code is a set of files which combine condition.py and find_neighbor.py. It's very difficult to work with right now, but more commits will have changes that iron out the many kinks.
main error currently is a stuck loop at iteration 22 - unclear so far what that is caused by. for the latest - run third.py